requirements.txt
asgiref==3.3.4
awesome-slugify==1.6.5
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.2
django-cors-headers==3.7.0
django-import-export==2.5.0
django-stdimage==5.3.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
flake8==3.9.2
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
pdfkit==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
sqlparse==0.4.1
tablib==3.0.0
Unidecode==0.4.21
xlrd==2.0.1
xlwt==1.3.0
